Basic UI building for filters.
The UI and the filter aren't connected right now, unfortunately.

Essentially the FilterUI takes a filter and introspects its
properties.  It then builds a UI for setting each of those
properties, and includes an 'execute' button for when the settings
are adequate.  When the execution finishes, the object emits an
'execution-success' signal, which the main application keys into.
A progress bar would be a nice addition...